requirements.txt
asgiref==3.11.1
black==26.1.0
blinker==1.9.0
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.4
crcmod==1.7
Deprecated==1.3.1
distlib==0.4.0
Django==6.0.2
djangorestframework==3.16.1
egauge-python==0.9.12
filelock==3.24.3
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.41
identify==2.6.16
idna==3.11
iniconfig==2.3.0
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
pre_commit==4.5.1
prometheus-python==1.0.3
prometheus_client==0.7.1
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
setuptools==82.0.0
smmap==5.0.2
sqlparse==0.5.5
urllib3==2.6.3
virtualenv==20.38.0
Werkzeug==3.1.5
wrapt==2.1.1
